feat: implement TerminalDisplay class for enhanced terminal UI; add cursor management and command handling
