+++ src/MainWindow.h
@@ -8,32 +8,9 @@
 #include <QString>
 #include <QTimer>
 
-class TerminalDisplay : public QPlainTextEdit {
-    Q_OBJECT
-public:
-    explicit TerminalDisplay(QWidget *parent = nullptr);
-    void appendOutput(const QString &text);
-    void setPrompt(const QString &prompt);
-    void updateCursor();
+#include "TerminalDisplay.h"
 
-signals:
-    void executeCommand(const QString &command);
-
-protected:
-    void paintEvent(QPaintEvent *event) override;
-    void keyPressEvent(QKeyEvent *event) override;
-    void resizeEvent(QResizeEvent *event) override;
-
-private:
-    QString m_prompt;
-    QString m_currentInput;
-    int m_cursorPos;
-    bool m_cursorVisible;
-    QTimer m_cursorTimer;
-};
-
-class MainWindow : public QWidget
-{
+class MainWindow : public QWidget {
     Q_OBJECT
 
 public:
