+++ src/main.cpp
@@ -1,16 +1,10 @@
 #include <QApplication>
 #include "MainWindow.h"
 
-int main(int argc, char *argv[]) {
+int main(int argc, char *argv[])
+{
     QApplication app(argc, argv);
-
-    MainWindow window;
-    window.show();
-
+    MainWindow w;
+    w.show();
     return app.exec();
 }
-// Compare this snippet from src/MainWindow.cpp:
-// #include "MainWindow.h"
-// #include <QPushButton>
-// #include <QVBoxLayout>
-// #include <QKeyEvent>
